refactor: moved and removed logic from helper userSchemas to user middleware, schema
